chore: refactor inputs into reusable components.
